Remove monkeypatch. Replace path funcs in pipeline. Remove session call in paths
