+++ workflow_calcium_imaging/__init__.py
@@ -1,0 +1,6 @@
+import datajoint as dj
+
+if "custom" not in dj.config:
+    dj.config["custom"] = {}
+
+db_prefix = dj.config["custom"].get("database.prefix", "")
