+++ workflow_calcium_imaging/pipeline.py
@@ -6,12 +6,16 @@ from element_event import trial, event
 from element_calcium_imaging import scan, imaging
 from element_lab.lab import Source, Lab, Protocol, User, Location, Project
 from element_animal.subject import Subject
+from .paths import (
+    get_imaging_root_data_dir,
+    get_scan_image_files,
+    get_scan_box_files,
+    get_nd2_files,
+    get_prairieview_files,
+)
 from . import analysis
-
-if "custom" not in dj.config:
-    dj.config["custom"] = {}
-
-db_prefix = dj.config["custom"].get("database.prefix", "")
+from .reference import Equipment
+from . import db_prefix
 
 __all__ = [
     "subject",
@@ -30,6 +34,11 @@ __all__ = [
     "Project",
     "Session",
     "Location",
+    "get_imaging_root_data_dir",
+    "get_scan_image_files",
+    "get_scan_box_files",
+    "get_nd2_files",
+    "get_prairieview_files",
 ]
 
 
@@ -47,23 +56,6 @@ session.activate(db_prefix + "session", linking_module=__name__)
 # Activate "event" and "trial" schema ---------------------------------
 
 trial.activate(db_prefix + "trial", db_prefix + "event", linking_module=__name__)
-
-
-# ------------- Declare table Equipment for use in element_calcium_imaging -------------
-
-
-@lab.schema
-class Equipment(dj.Manual):
-    """Equipment
-
-    Attributes:
-        scanner (str): Scanner used in imaging.
-    """
-
-    definition = """
-    scanner: varchar(32)
-    """
-
 
 # ------------- Activate "imaging" schema -------------
