+++ workflow_calcium_imaging/reference.py
@@ -1,0 +1,17 @@
+import datajoint as dj
+from . import db_prefix
+
+schema = dj.Schema(db_prefix + "reference")
+
+
+@schema
+class Equipment(dj.Manual):
+    """Equipment
+
+    Attributes:
+        scanner (str): Scanner used in imaging.
+    """
+
+    definition = """
+    scanner: varchar(32)
+    """
